Reduced code in audio.js in preparation to actually add or remove nodes dynamically to the graph.  Broke out code into new directories. Removed unnecessary html and event listeners
