+++ public/index.html
@@ -30,7 +30,6 @@
       <canvas id="visualizer"></canvas>
       <h2>Play some sounds!</h2>
       <button id="create">Create</button>
-      <button id="modify">Modify</button>
       <form>
         <div>
           <label for="volume">
@@ -41,19 +40,6 @@
               min="0"
               max="10"
               value="5"
-              step="0.01"
-            />
-          </label>
-        </div>
-        <div>
-          <label for="distortion">
-            Distortion
-            <input
-              type="range"
-              id="distortion"
-              min="0"
-              max="10"
-              value="1"
               step="0.01"
             />
           </label>
