+++ public/js/audio.js
@@ -1,10 +1,8 @@
-// import DOM from './dom/dom.js';
+import DOM from './dom/dom.js';
 import * as Effects from './effects/index';
 import createTrack from './track.js';
+import { createTrackGraph } from './graph/track';
 import waveformVisualizer from './dom/visualizer.js';
-
-const createGraph = (nodes, track) =>
-  nodes.reduce((acc, node) => acc.connect(node), track);
 
 export default async function initialize() {
   const audioContext = new AudioContext();
@@ -12,55 +10,27 @@ export default async function initialize() {
 
   const track = createTrack(audioContext, audioElement);
 
-  // const convolver = await impulseResponse(audioContext);
   const analyser = audioContext.createAnalyser();
 
-  // const delayNode = Delay(audioContext);
-  // const distortionNode = Distortion(audioContext);
-  const { Fuzz } = Effects;
-  const fuzzNodes = Fuzz(audioContext, track);
+  const { Fuzz, Gain } = Effects;
 
-  // const preAmpGainNode = Gain(audioContext);
-  // const postAmpGainNode = Gain(audioContext, { gain: 2 });
-  // const reverbNode = await Reverb(audioContext);
-  // const compressorNode = await Compressor(audioContext, {
-  //   threshold: -40,
-  //   knee: 30,
-  //   attack: 1,
-  //   release: 1,
-  //   ratio: 15,
-  // });
-
+  // TODO: nodes should be an array of wrapped nodes.
+  // Each node will contain the audio api object and metadata.
   const nodes = [
-    // preAmpGainNode,
-    ...fuzzNodes,
-    // postAmpGainNode,
+    Gain(audioContext),
+    ...Fuzz(audioContext, track),
     analyser,
     audioContext.destination,
   ];
 
   waveformVisualizer(analyser);
 
-  const createButton = document.getElementById('create');
-  createButton.addEventListener('click', () => {
-    createGraph(nodes, track);
-
-    if (audioContext.state === 'suspended') {
-      audioContext.resume();
-    }
-    audioElement.play();
+  DOM({
+    nodes,
+    createTrackGraph: () => createTrackGraph(nodes, track),
+    audioContext,
+    audioElement,
   });
-
-  // const modifyButton = document.getElementById('modify');
-  // modifyButton.addEventListener('click', () => {
-  //   const newNodes = [...nodes.slice(2)];
-
-  //   nodes.reverse().forEach((node) => node.disconnect());
-
-  //   createGraph(newNodes, track);
-  // });
-
-  // DOM({ gainNode: preAmpGainNode, distortionNode }, audioContext);
 }
 
 initialize();
